# BACKEND/openapi.yaml
openapi: 3.1.0
info:
  title: SIG-DESK API
  version: 0.1.0-beta
servers:
  - url: http://localhost:8080/api/v1
paths:
  /tickets:
    get:
      operationId: listTickets
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/TicketStatus" }
        - name: priority
          in: query
          schema: { $ref: "#/components/schemas/TicketPriority" }
        - name: category
          in: query
          schema: { type: string }
        - name: site
          in: query
          schema: { type: string }
        - name: assignee
          in: query
          schema: { type: string }
        - name: unassigned
          in: query
          schema: { type: boolean }
        - name: q
          in: query
          description: Free-text search over title, description and human id.
          schema: { type: string }
        - name: cursor
          in: query
          description: Opaque keyset-pagination cursor from a previous response's nextCursor.
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: mergedInto
          in: query
          description: Overrides the default (exclude merged tickets) and returns only tickets absorbed into this human id.
          schema: { type: string }
      responses:
        "200":
          description: Ticket page
          content:
            application/json:
              schema:
                type: object
                required: [items, hasMore]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Ticket"
                  nextCursor:
                    type: string
                  hasMore:
                    type: boolean
    post:
      operationId: createTicket
      deprecated: true
      description: Legacy creation path. New INC records use POST /entities/INC.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicket"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    parameters:
      - $ref: "#/components/parameters/TicketID"
    get:
      operationId: getTicket
      responses:
        "200":
          description: Ticket detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Ticket not found
  /tickets/{id}/status:
    parameters:
      - $ref: "#/components/parameters/TicketID"
    patch:
      operationId: updateTicketStatus
      deprecated: true
      description: Legacy transition path. New INC records use the metadata runtime transition endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/TicketStatus"
                actorName:
                  type: [string, "null"]
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "409":
          description: Invalid status transition
  /tickets/{id}/assign:
    parameters:
      - $ref: "#/components/parameters/TicketID"
    post:
      operationId: assignTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                assigneeName:
                  type: [string, "null"]
                  description: Null unassigns the ticket.
                actorName:
                  type: [string, "null"]
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/merge:
    parameters:
      - $ref: "#/components/parameters/TicketID"
    post:
      operationId: mergeTickets
      description: Merges other tickets into this one; the merged tickets are marked resolved and linked via mergedIntoId.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mergedIds]
              properties:
                mergedIds:
                  type: array
                  items: { type: string }
                actorName:
                  type: [string, "null"]
      responses:
        "200":
          description: Primary ticket after merge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/unmerge/{mergedId}:
    parameters:
      - $ref: "#/components/parameters/TicketID"
      - name: mergedId
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: unmergeTicket
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                actorName:
                  type: [string, "null"]
      responses:
        "200":
          description: Primary ticket after unmerge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/comments:
    parameters:
      - $ref: "#/components/parameters/TicketID"
    get:
      operationId: listTicketComments
      responses:
        "200":
          description: Comments in chronological order
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Comment" }
    post:
      operationId: addTicketComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                authorName: { type: string }
                body: { type: string }
                isInternal: { type: boolean, default: false }
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" }
  /tickets/{id}/attachments:
    parameters:
      - $ref: "#/components/parameters/TicketID"
    get:
      operationId: listTicketAttachments
      responses:
        "200":
          description: Attachments in chronological order
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Attachment" }
    post:
      operationId: uploadTicketAttachment
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                uploaderName: { type: string }
      responses:
        "201":
          description: Attachment stored
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attachment" }
  /attachments/{attachmentId}/download:
    parameters:
      - name: attachmentId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: downloadTicketAttachment
      responses:
        "200":
          description: Raw file bytes
          content:
            application/octet-stream: {}
        "404":
          description: Attachment not found
  /tickets/{id}/watchers:
    parameters:
      - $ref: "#/components/parameters/TicketID"
    get:
      operationId: listTicketWatchers
      responses:
        "200":
          description: Watchers
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Watcher" }
    post:
      operationId: addTicketWatcher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [watcherName]
              properties:
                watcherName: { type: string }
      responses:
        "204":
          description: Watcher added
  /tickets/{id}/watchers/{watcherName}:
    parameters:
      - $ref: "#/components/parameters/TicketID"
      - name: watcherName
        in: path
        required: true
        schema: { type: string }
    delete:
      operationId: removeTicketWatcher
      responses:
        "204":
          description: Watcher removed
  /tickets/{id}/activity:
    parameters:
      - $ref: "#/components/parameters/TicketID"
    get:
      operationId: listTicketActivity
      responses:
        "200":
          description: Activity stream in chronological order
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ActivityEntry" }
  /catalog/layouts/{entityKey}/draft:
    parameters:
      - $ref: "#/components/parameters/EntityKey"
    get:
      operationId: getCatalogLayoutDraft
      responses:
        "200":
          description: Active layout draft
        "404":
          description: Draft not found
    post:
      operationId: createCatalogLayoutDraft
      responses:
        "201":
          description: Draft created
        "409":
          description: Draft already exists
    put:
      operationId: updateCatalogLayoutDraft
      responses:
        "200":
          description: Draft updated
        "404":
          description: Draft not found
  /catalog/layouts/{entityKey}/publish:
    parameters:
      - $ref: "#/components/parameters/EntityKey"
    post:
      operationId: publishCatalogLayoutDraft
      responses:
        "200":
          description: Layout published
        "422":
          description: Incompatible layout document
  /catalog/layouts/{entityKey}/versions:
    parameters:
      - $ref: "#/components/parameters/EntityKey"
    get:
      operationId: listCatalogLayoutVersions
      responses:
        "200":
          description: Layout version history
  /catalog/layouts/{entityKey}/active:
    parameters:
      - $ref: "#/components/parameters/EntityKey"
    get:
      operationId: getActiveCatalogLayoutVersion
      responses:
        "200":
          description: Active published layout version
        "404":
          description: Active layout version not found
  /catalog/layouts/{entityKey}/versions/{version}/activate:
    parameters:
      - $ref: "#/components/parameters/EntityKey"
      - name: version
        in: path
        required: true
        schema: { type: integer }
    post:
      operationId: activateCatalogLayoutVersion
      responses:
        "200":
          description: Version activated
        "422":
          description: Target version is not published
  /entities/{entityKey}/{entityID}/resolved-definition:
    parameters:
      - $ref: "#/components/parameters/EntityKey"
      - name: entityID
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getResolvedDefinitionForRecord
      responses:
        "200":
          description: Resolved definition and layout for target record
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Record not found
  /catalog/definitions:
    get:
      operationId: listCatalogDefinitions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [published]
      responses:
        "200":
          description: Definition versions
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/CatalogDefinition" }
    post:
      operationId: createCatalogDefinitionDraft
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CatalogDefinitionDraft" }
      responses:
        "201":
          description: Immutable draft version created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogDefinition" }
        "422":
          description: Invalid metamodel
  /catalog/definitions/{entityKey}:
    parameters:
      - $ref: "#/components/parameters/EntityKey"
    get:
      operationId: getPublishedCatalogDefinition
      responses:
        "200":
          description: Current published executable definition
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogDefinition" }
        "404":
          description: No published definition
  /catalog/definitions/{entityKey}/versions/{version}/validate:
    parameters:
      - $ref: "#/components/parameters/EntityKey"
      - $ref: "#/components/parameters/DefinitionVersion"
    post:
      operationId: validateCatalogDefinitionPublication
      description: Resolves module-owned resources and validates version compatibility.
      responses:
        "200":
          description: Publication validation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicationValidation" }
  /catalog/definitions/{entityKey}/versions/{version}/publish:
    parameters:
      - $ref: "#/components/parameters/EntityKey"
      - $ref: "#/components/parameters/DefinitionVersion"
    post:
      operationId: publishCatalogDefinition
      description: Compiles and persists an immutable, version-locked executable manifest.
      responses:
        "200":
          description: Published definition
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogDefinition" }
        "422":
          description: Publication validation failed
  /catalog/definitions/{entityKey}/versions/{version}/manifest:
    parameters:
      - $ref: "#/components/parameters/EntityKey"
      - $ref: "#/components/parameters/DefinitionVersion"
    get:
      operationId: getExecutableDefinitionManifest
      responses:
        "200":
          description: Immutable executable manifest
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExecutableDefinitionManifest" }
...
  /entities/{entityKey}/{entityID}/transitions/{transitionKey}:
    parameters:
      - $ref: "#/components/parameters/EntityKey"
      - name: entityID
        in: path
        required: true
        schema: { type: string }
      - name: transitionKey
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: transitionEntityRecord
      responses:
        "200":
          description: Record after interpreting the transition
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EntityRecord" }
        "409":
          description: Concurrent state change
        "422":
          description: Transition is not available from the current state
...
components:
  parameters:
...
    TicketID:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: "^INC-[0-9]{6}$"
  schemas:
...
    TicketStatus:
      type: string
      enum: [open, in_progress, pending_review, resolved]
    TicketPriority:
      type: string
      enum: [low, medium, high, critical]
    CreateTicket:
      type: object
      required: [title, description, priority, category]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 160
        description:
          type: string
          minLength: 10
          maxLength: 10000
        priority:
          $ref: "#/components/schemas/TicketPriority"
        category:
          type: string
        requesterName:
          type: string
        assetId:
          type: [string, "null"]
        site:
          type: [string, "null"]
    Ticket:
      allOf:
        - $ref: "#/components/schemas/CreateTicket"
        - type: object
          required: [id, status, requesterName, createdAt, mergedCount]
          properties:
            id:
              type: string
            status:
              $ref: "#/components/schemas/TicketStatus"
            requesterName:
              type: string
            assigneeName:
              type: [string, "null"]
            createdAt:
              type: string
              format: date-time
            mergedCount:
              type: integer
              minimum: 0
            mergedIntoId:
              type: [string, "null"]
              description: Set when this ticket was absorbed by another ticket via merge.
...
    Watcher:
      type: object
      required: [ticketId, watcherName, createdAt]
      properties:
        ticketId: { type: string }
        watcherName: { type: string }
        createdAt: { type: string, format: date-time }
...